refactor: move fcm.rs to alc-core, health.rs to alc-misc
Move FcmSender implementation into alc-core/src/fcm.rs and health
check route into alc-misc/src/health.rs. src/ now contains only
entry points and re-export facades (523 lines total).
