+++ crates/alc-misc/src/health.rs
@@ -1,7 +1,7 @@
 use axum::{routing::get, Json, Router};
 use serde_json::{json, Value};
 
-use crate::AppState;
+use alc_core::AppState;
 
 pub fn router() -> Router<AppState> {
     Router::new().route("/health", get(health_check))
