+++ crates/alc-misc/src/lib.rs
@@ -5,6 +5,7 @@ pub mod communication_items;
 pub mod driver_info;
 pub mod employees;
 pub mod guidance_records;
+pub mod health;
 pub mod measurements;
 pub mod repo;
 pub mod sso_admin;
