+++ src/routes/mod.rs
@@ -2,6 +2,7 @@
 pub use alc_carins::car_inspection_files;
 pub use alc_carins::car_inspections;
 pub use alc_carins::carins_files;
+pub use alc_carins::nfc_tags;
 pub use alc_devices::devices;
 pub use alc_dtako::dtako_csv_proxy;
 pub use alc_dtako::dtako_daily_hours;
@@ -21,15 +22,14 @@ pub use alc_misc::communication_items;
 pub use alc_misc::driver_info;
 pub use alc_misc::employees;
 pub use alc_misc::guidance_records;
-pub use alc_tenko::daily_health;
-pub use alc_tenko::equipment_failures;
-pub mod health;
-pub use alc_carins::nfc_tags;
+pub use alc_misc::health;
 pub use alc_misc::measurements;
 pub use alc_misc::sso_admin;
 pub use alc_misc::tenant_users;
 pub use alc_misc::timecard;
 pub use alc_misc::upload;
+pub use alc_tenko::daily_health;
+pub use alc_tenko::equipment_failures;
 pub use alc_tenko::health_baselines;
 pub use alc_tenko::tenko_call;
 pub use alc_tenko::tenko_records;
